+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 3.1.0 / Upcoming
 ================
 
-* Add (or remove) items in the catalog.
+* Add and remove catalogs.
+* Add and remove items in the catalog.
 * Update deps.
 
 3.0.2 / 2016-05-30